requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
setuptools>=65.0.0
google-cloud-vision==3.3.0
google-cloud-storage==2.10.0
Pillow==10.0.0